Move library\OAuth1.php to class structure Friendica\Security\OAuth1
