+++ src/Security/FKOAuth1.php
@@ -19,15 +19,14 @@
  *
  */
 
-namespace Friendica\Network;
+namespace Friendica\Security;
 
 use Friendica\Core\Logger;
 use Friendica\Database\DBA;
 use Friendica\DI;
-use Friendica\Security\FKOAuthDataStore;
-use OAuthServer;
-use OAuthSignatureMethod_HMAC_SHA1;
-use OAuthSignatureMethod_PLAINTEXT;
+use Friendica\Security\OAuth1\OAuthServer;
+use Friendica\Security\OAuth1\OAuthSignatureMethod_HMAC_SHA1;
+use Friendica\Security\OAuth1\OAuthSignatureMethod_PLAINTEXT;
 
 /**
  * OAuth protocol
