+++ src/Security/OAuth1/OAuthException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Friendica\Security\OAuth1;
+
+use Exception;
+
+class OAuthException extends Exception
+{
+}
