+++ src/Security/OAuth1/README.md
@@ -1,0 +1,1 @@
+This namespace contains the OAuth1 library for server and client usages
